third_party/blink/web_tests/platform/generic/html/details_summary/details-keyboard-show-hide-expected.txt

This test verifies that pressing the ENTER or SPACEBAR key when the <summary> has focus will toggle the <details> display. On expanding details.open will be true and vice versa.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS openAttribute("details") is false
Toggle <display> using Enter key:
PASS openAttribute("details") is true
PASS openAttribute("details") is false
Toggle <display> using Spacebar key:
PASS openAttribute("details") is true
PASS openAttribute("details") is false
PASS successfullyParsed is true

TEST COMPLETE
